pyproject.toml
[project]
name = "fd-g-flowmeter-testing"
version = "0.1.0"
description = "FD-G FlowMeter ModbusTCP Communication Testing Tool"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyModbusTCP>=0.2.0",
    'paho-mqtt>=2.1.0',
    "python-dotenv>=1.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
fdg-test = "fdg_flowmeter:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.black]
line-length = 88
target-version = ['py38']